+++ app/Http/Controllers/TransactionController.php
@@ -31,20 +31,25 @@ class TransactionController extends Controller
      */
     public function store(Request $request)
     {
-        $ticket = Ticket::find($request->id);
-        $ticket_price = $ticket->price;
-        $total_price = $request->amount_of_ticket * $ticket_price;
-        if($total_price != 0){
-            $tax = $total_price * (1/100);
+        $data = $request->validate([
+            'ticket_id' => 'required|integer',
+            'amount_of_ticket' => 'required|integer',
+            'role' => 'required|in:ARTISTS,DEVS,USERS'
+        ]);
+
+        if($data['role'] === "ARTISTS" || $data['role'] === "DEVS"){
+            $ticket = Ticket::find($data['ticket_id']);
+            $ticket_price = $data['amount_of_ticket'] * $ticket->price;
+            $tax = $ticket_price * (11/100);
+            $total_price = $ticket_price + $tax;
             $transactions = Transaction::create([
                 'transaction_id' => uniqid(),
                 'amount_of_ticket' => $request->amount_of_ticket,
                 'ticket_price' => $ticket_price,
                 'total_price' => $total_price,
                 'concert_name' => $ticket->concert_name,
-                'concert_address' => $ticket->address,
+                'concert_address' => $ticket->concert_address,
                 'concert_date' => $ticket->concert_date,
-                'tax' => $tax,
                 'currency' => $ticket->currency
             ]);
             return response()->json([
@@ -54,12 +59,11 @@ class TransactionController extends Controller
             ], 200);
         } else {
             return response()->json([
-                'status' => 406,
-                'message' => "total_price can't be 0",
-                'data' => 'null'
-            ], 406);
+                'status' => 405,
+                'message' => "permission denied",
+                'data' => "$request->role don't have permission to this method"
+            ], 405);
         }
-
     }
 
     /**
@@ -80,8 +84,8 @@ class TransactionController extends Controller
         } else {
             return response()->json([
                 'status' => 404,
-                'message' => "transaction id $id  not found",
-                'data' => 'null'
+                'message' => "transaction not found",
+                'data' => "transaction with $id not found"
             ], 404);
         };
     }
@@ -95,49 +99,47 @@ class TransactionController extends Controller
      */
     public function update(Request $request, $id)
     {
-        if($request->id != null){
+        $data = $request->validate([
+            'ticket_id' => 'required|integer',
+            'amount_of_ticket' => 'required|integer',
+            'role' => 'required|in:ARTISTS,DEVS,USERS'
+        ]);
+
+        
+        if($data['role'] === "ARTISTS" || $data['role'] === "DEVS"){
             $transactions = Transaction::find($id);
-            $ticket = Ticket::find($request->id);
-            $ticket_price = $ticket->price;
-            $total_price = $request->amount_of_ticket * $ticket_price;
-            $tax = $total_price * (1/100);
+            $ticket = Ticket::find($data['ticket_id']);
+            $ticket_price = $data['amount_of_ticket'] * $ticket->price;
+            $tax = $ticket_price * (11/100);
+            $total_price = $ticket_price + $tax;
             if($transactions){
-                if($request->amount_of_ticket != 0){
-                    $transactions->transaction_id = $transactions->transaction_id;
-                    $transactions->total_price = $total_price;
-                    $transactions->amount_of_ticket = $request->amount_of_ticket;
-                    $transactions->ticket_price = $ticket_price;
-                    $transactions->concert_name = $ticket->concert_name;
-                    $transactions->concert_address = $ticket->address;
-                    $transactions->concert_date = $ticket->concert_date;
-                    $transactions->tax = $tax;
-                    $transactions->currency = $ticket->currency;
-                    $transactions->save();
-                    return response()->json([
-                        'status' => 200,
-                        'message' => 'data successfully updated',
-                        'data' => $transactions
-                    ], 200);
-                } else {
-                    return response()->json([
-                        'status' => 418,
-                        'message' => "amount_of_id can't be 0", 
-                        'data' => 'null'
-                    ], 418);
-                }
+                $transactions->transaction_id = $transactions->transaction_id;
+                $transactions->total_price = $total_price;
+                $transactions->amount_of_ticket = $request->amount_of_ticket;
+                $transactions->ticket_price = $ticket_price;
+                $transactions->concert_name = $ticket->concert_name;
+                $transactions->concert_address = $ticket->concert_address;
+                $transactions->concert_date = $ticket->concert_date;
+                $transactions->currency = $ticket->currency;
+                $transactions->save();
+                return response()->json([
+                    'status' => 200,
+                    'message' => 'data successfully updated',
+                    'data' => $transactions
+                ], 200);
             } else {
                 return response()->json([
                     'status' => 404,
-                    'message' => "transaction id $id not found", 
-                    'data' => 'null'
+                    'message' => "transaction not found", 
+                    'data' => "transaction with $id not found",
                 ], 404);
-            };
+            };   
         } else {
             return response()->json([
-                'status' => 418,
-                'message' => "ticket_id is required", 
-                'data' => 'null'
-            ], 418);
+                'status' => 405,
+                'message' => "permission denied",
+                'data' => "$request->role don't have permission to this method"
+            ], 405);
         }
     }
 
@@ -147,22 +149,36 @@ class TransactionController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy($id, Request $request)
     {
-        $transactions = Transaction::where('id', $id)->first();
-        if($transactions){
-            $transactions->delete();
-            return response()->json([
-                'status' => 200,
-                'message' => "data successfully deleted",
-                'data' => $transactions
-            ], 200);
+        $data = $request->validate([
+            'ticket_id' => 'required|integer',
+            'amount_of_ticket' => 'required|integer',
+            'role' => 'required|in:ARTISTS,DEVS,USERS'
+        ]);
+
+        if($data['role'] === "ARTISTS" || $data['role'] === "DEVS"){
+            $transactions = Transaction::where('id', $id)->first();
+            if($transactions){
+                $transactions->delete();
+                return response()->json([
+                    'status' => 200,
+                    'message' => "data successfully deleted",
+                    'data' => $transactions
+                ], 200);
+            } else {
+                return response()->json([
+                    'status' => 404,
+                    'message' => "transaction not found",
+                    'data' => "transaction with $id not found"
+                ], 404);
+            }   
         } else {
             return response()->json([
-                'status' => 404,
-                'message' => "transaction id $id not found",
-                'data' => 'null'
-            ], 404);
+                'status' => 405,
+                'message' => "permission denied",
+                'data' => "$request->role don't have permission to this method"
+            ], 405);
         }
     }
 }
